LayoutTests/storage/indexeddb/modern/objectstore-rename-abort-private-expected.txt

This test verifies renaming object store can be reverted succesfully if version change transaction aborts.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


openRequest = indexedDB.open(dbname);
transaction = openRequest.transaction;
database = openRequest.result;
objectStore = database.createObjectStore('os1');
database.deleteObjectStore('os1');
objectStore2 = database.createObjectStore('os2')
objectStore2.name = 'os1';
objectStore2.put(1, 1).onsuccess = () => { transaction.abort(); }
newOpenRequest = indexedDB.open(dbname);
newDatabase = newOpenRequest.result;
PASS newDatabase.objectStoreNames.length is 0
PASS successfullyParsed is true

TEST COMPLETE
